pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "mymoney"
version = "1.0"
description = "Visualize and keep track of personal finances"
authors = [
	{name = "Gary Faircloth", email = "arkadianriver@gmail.com"}
]
requires-python = ">=3.9"
readme = "README.md"
classifiers = [
	"License :: OSI Approved :: MIT License",
]
dependencies = [
	"numpy>=1.26.2",
	"packaging>=23.2",
	"pandas>=2.1.3",
	"plotly>=5.18.0",
	"python-dateutil>=2.8.2",
	"PyYAML>=5.4.1",
]

[project.urls]
Homepage = "https://github.com/arkadianriver/mymoney"

[project.scripts]
mymoney = 'mymoney.mymoney:main'